+++ simple_history/tests/tests.py
@@ -6,6 +6,9 @@ from django.test import TestCase
 from django_webtest import WebTest
 from django.core.files.base import ContentFile
 from django.core.urlresolvers import reverse
+from simple_history.tests.models import Profile, AdminProfile, Bookcase
+from django.db import models
+from simple_history.models import HistoricalRecords
 try:
     from django.contrib.auth import get_user_model
     User = get_user_model()
@@ -219,7 +222,6 @@ class HistoricalRecordsTest(TestCase):
             'history_type': "~",
         })
 
-
 class RegisterTest(TestCase):
     def test_register_no_args(self):
         self.assertEqual(len(Choice.history.all()), 0)
@@ -243,6 +245,25 @@ class RegisterTest(TestCase):
         self.assertTrue(hasattr(User, 'histories'))
         self.assertFalse(hasattr(User, 'again'))
 
+class CreateHistoryModelTests(TestCase):
+
+    def test_create_history_model_with_one_to_one_field_to_integer_field(self):
+        records = HistoricalRecords()
+        records.module = AdminProfile.__module__
+        try:
+            records.create_history_model(AdminProfile)
+        except:
+            self.fail("SimpleHistory should handle foreign keys to one to one"
+                      "fields to integer fields without throwing an exception.")
+    
+    def test_create_history_model_with_one_to_one_field_to_char_field(self):
+        records = HistoricalRecords()
+        records.module = Bookcase.__module__
+        try:
+            records.create_history_model(Bookcase)
+        except:
+            self.fail("SimpleHistory should handle foreign keys to one to one"
+                      "fields to char fields without throwing an exception.")
 
 class AppLabelTest(TestCase):
     def get_table_name(self, manager):